Extract type conversions tests to separate file
